Split Eventing tests into unit and integration projects
