+++ src/EventLogExpert.Eventing.IntegrationTests/EventLogExpert.Eventing.IntegrationTests.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <IsPackable>false</IsPackable>
+    <IsTestProject>true</IsTestProject>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\EventLogExpert.Eventing\EventLogExpert.Eventing.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="..\EventLogExpert.Eventing.Tests\TestUtils\Constants\Constants.Provider.cs" Link="TestUtils\Constants\Constants.Provider.cs" />
+  </ItemGroup>
+
+</Project>
