+++ src/EventLogExpert.Eventing.IntegrationTests/Providers/RegistryProviderTests.cs
@@ -6,7 +6,7 @@ using EventLogExpert.Eventing.Providers;
 using EventLogExpert.Eventing.Tests.TestUtils.Constants;
 using NSubstitute;
 
-namespace EventLogExpert.Eventing.Tests.Providers;
+namespace EventLogExpert.Eventing.IntegrationTests.Providers;
 
 public sealed class RegistryProviderTests
 {
